requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
colorama==0.4.4
coverage==5.5
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
ujson==4.0.2
Werkzeug==1.0.1
